Just finished adding everything again, Servo support is offline for the moment. This is because the stock firmware ignores the second servo port because it can sometimes interfere with the operation of an external extruderboard stepper motor. This is why compiler args were invented! Now just need a Makerbot cupcake to test on....
Changed the ZProbe servo pin from D10 to D9.
This was because pin D10 was disabled in the Makerbot firmware to make way for external stepper support on the extruder controller.
ZProbe support has been re-enabled.

Forgot to add the EEPROM32 command. Also dealt with some 5Axis Problems.
(I can't believe that the stepper extruders are treated as Axes!) Works now for Thingomatics without stepper extruders. Need now to test with a stepper extruder thingomatic.

After homing, the makerbot now disables its steppers. For safety!

This code is now tested and in working order. (Tested on a Thingomatic with a DC extruder. For the sake of testing, this should be retested on a makerbot with a stepper extruder.)